requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=py38h06a4308_0
aiohttp=3.8.3=py38h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py38h06a4308_0
attrs=22.1.0=py38h06a4308_0
backcall=0.2.0=pyhd3eb1b0_0
binutils_impl_linux-64=2.38=h2a08ee3_1
binutils_linux-64=2.38.0=hc2dff05_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
bottleneck=1.3.5=py38h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py38h27cfd23_1003
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
clingo=5.6.2=py38h3fd9d12_0
contourpy=1.0.5=py38hdb19cb5_0
cryptography=38.0.1=py38h9ce1e76_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=h6a678d5_0
fftw=3.3.9=h27cfd23_1
flit-core=3.6.0=pyhd3eb1b0_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py38h5eee18b_0
gcc_impl_linux-64=11.2.0=h1234567_1
gcc_linux-64=11.2.0=h5c386dc_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.18=pyhd3eb1b0_1
glib=2.69.1=he621ea3_2
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
grpcio=1.42.0=py38hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gxx_impl_linux-64=11.2.0=h1234567_1
gxx_linux-64=11.2.0=hc2dff05_0
h5py=3.7.0=py38h737f45e_0
hdf5=1.10.6=h3ffc7dd_1
icu=58.2=he6710b0_3
idna=3.4=py38h06a4308_0
importlib-metadata=4.11.3=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipython=8.7.0=py38h06a4308_0
jedi=0.18.1=py38h06a4308_1
jpeg=9e=h7f8727e_0
kernel-headers_linux-64=2.6.32=he073ed8_15
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-devel_linux-64=11.2.0=h1234567_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libstdcxx-devel_linux-64=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_2
libuuid=1.41.5=h5eee18b_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py38h06a4308_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.6.2=py38h06a4308_0
matplotlib-base=3.6.2=py38h945d387_0
matplotlib-inline=0.1.6=py38h06a4308_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=6.0.2=py38h5eee18b_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=py38he184ba9_0
numpy=1.23.4=py38h14f4228_0
numpy-base=1.23.4=py38h31eccc5_0
oauthlib=3.2.1=py38h06a4308_0
openssl=1.1.1s=h7f8727e_0
packaging=22.0=py38h06a4308_0
pandas=1.5.2=py38h417a72b_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py38hace64e9_1
pip=22.3.1=py38h06a4308_0
ply=3.11=py38_0
prompt-toolkit=3.0.36=py38h06a4308_0
protobuf=3.20.1=py38h295c915_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=py38h06a4308_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5-sip=12.11.0=py38h6a678d5_1
pysocks=1.7.1=py38h06a4308_0
python=3.8.15=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pytz=2022.7=py38h06a4308_0
pyyaml=6.0=py38h5eee18b_1
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.1=py38h06a4308_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scipy=1.9.3=py38h14f4228_0
seaborn=0.12.1=py38h06a4308_0
setuptools=65.5.0=py38h06a4308_0
sip=6.6.2=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
sqlite=3.40.0=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
sysroot_linux-64=2.12=he073ed8_15
tensorboard=2.10.0=py38h06a4308_0
tensorboard-data-server=0.6.1=py38h52d8a92_0
tensorboard-plugin-wit=1.8.1=py38h06a4308_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.7.2=py38
torchvision=0.13.1=cpu_py38h164cc8f_0
tornado=6.2=py38h5eee18b_0
tqdm=4.64.1=py38h06a4308_0
traitlets=5.7.1=py38h06a4308_0
typing-extensions=4.4.0=py38h06a4308_0
typing_extensions=4.4.0=py38h06a4308_0
urllib3=1.26.13=py38h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.2=py38h06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py38h5eee18b_0
zipp=3.11.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
